Migrate to M5Unified and ESP32 core 3.x (ESP-IDF 5.5)
M5Stack ライブラリ 0.4.6 は 2023 年で更新が止まっており、ESP-IDF 5.x で
削除された rom/miniz.h に依存するため core 3.x でビルドできない (README の
Notes にある「board manager を 2.1.4 に下げる」制約)。デバイス層を M5Unified
に移行し、m5stack:esp32@3.3.8 (ESP-IDF v5.5.4) でビルドする。

主な変更:

- IMU は FIFO の直接バースト読み出しから M5.Imu.update() のポーリングへ。
  Madgwick には imu_data.usec から実測した積分幅を渡す
  (MadgwickAHRSetDeltaT を追加)
- ジャイロオフセットを生 ADC から deg/s に変更。型が変わるため NVS キーも
  gryo_* から gyro_* に変更しており、キャリブレーションのやり直しが必要
- BALA2 ベースの I2C を M5.In_I2C.writeRegister/readRegister へ
- isChargeFull() は M5Unified に無いため getBatteryLevel() >= 100 で代替
- M5Unified はボタン状態を update() で確定するため、起動時のボタン判定前に
  M5.update() を呼ぶ
- Arduino CLI のビルドプロファイル sketch.yaml を追加
- スケッチ名をリポジトリ名に合わせて変更 (Arduino CLI はディレクトリ名と
  .ino 名の一致を要求するため)

MPU6886 の加速度・ジャイロについて、M5Unified は M5Stack Fire で軸の入れ替えを
行わず (反転は地磁気のみ)、単位も g / deg/s で従来の生値換算と一致する。

+++ src/calibration.h
@@ -3,4 +3,6 @@
 void calibrationGryo();
 void calibrationInit();
 void calibrationSaveCenterAngle(float angle);
-void calibrationGet(int16_t* gyro_x_offset, int16_t* gyro_y_offset, int16_t* gyro_z_offset, float *angle_center);
+
+// Gyro offsets are in deg/s, matching the units M5Unified reports.
+void calibrationGet(float* gyro_x_offset, float* gyro_y_offset, float* gyro_z_offset, float *angle_center);
